% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processOutput.R
\name{processOutput}
\alias{processOutput}
\title{implement this function to handle the GS output}
\usage{
processOutput(msg, io, saveConfig = list(enabled = FALSE))
}
\description{
implement this function to handle the GS output
}
