% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripGenVerbatim.R
\name{stripGenVerbatim}
\alias{stripGenVerbatim}
\title{Function to strip stuff that looks like this:<SPAN CLASS=GenVerbatim><PRE></PRE></SPAN>
from the output.}
\usage{
stripGenVerbatim(output)
}
\description{
Function to strip stuff that looks like this:<SPAN CLASS=GenVerbatim><PRE></PRE></SPAN>
from the output.
}
